requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
backports.print-function==1.1.1
blis==0.4.1
CacheMan==2.1.0
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextualSpellCheck==0.1.1
cycler==0.10.0
cymem==2.0.3
distlib==0.3.1
Django==3.0.8
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-2.3.1/en_core_web_md-2.3.1.tar.gz
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
filelock==3.0.12
Flask==1.1.2
freeze==3.0
future==0.18.2
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-vision==1.0.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
idna==2.10
image==1.5.32
importlib-metadata==1.7.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
murmurhash==1.0.2
ner==0.1
nltk==3.5
numpy==1.19.0
opencv-python==4.3.0.36
pandas==1.0.5
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
protobuf==3.12.2
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pylint==2.5.3
pyparsing==2.4.7
pyspellchecker==0.5.4
pytesseract==0.3.4
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.6.8
requests==2.24.0
rsa==4.6
six==1.15.0
spacy==2.3.1
spacy-symspell==0.1.2
sqlparse==0.3.1
srsly==1.0.2
symspellpy==6.5.2
thinc==7.4.1
tinydb==4.1.1
toml==0.10.1
tqdm==4.46.1
typed-ast==1.4.1
unicode-string-literal==1.1
urllib3==1.25.9
virtualenv==20.0.26
waitress==1.4.4
wasabi==0.7.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
